ml_server/requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
pillow
fastapi
uvicorn
python-multipart
python-dotenv